requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.115.12
h11==0.16.0
idna==3.10
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
langchain
langchain-community
langsmith
vllm
transformers
sentence-transformers
faiss-cpu
google-cloud-storage
python-dotenv>=1.0.0
arrow==1.3.0 
holidays==0.46
accelerate
torch
rank_bm25>=0.2.2
pytest
pytest-asyncio
httpx
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-system-metrics
psutil